Use server-side generated ID instead of volunteer fingerprint.
It turns out per-volunteer statistics or optimizations will not be implemented
in the near future and volunteer fingerprinting + ip address mixing as a way to
receive unique ID turns out to be too complicated and not necessary at the moment.
Instead of providing its own fingerprint volunteer now gets id from the server
and uses it in subsequent calls.